Rename `graphics::image` to disambiguate
Also clean up some examples and such.
